added LabeledFieldArray; added to Step3InviteTeam
